requirements.txt
aiofiles
telethon
google-generativeai
faiss-cpu
numpy
pypdf